+++ logic_gates_snn.py
